(platform-generic-c) Manage memory allocation for algebraic data types
The management of memory allocation for algebraic data types emulates the behaviour of a stack-based allocation policy. Whenever an algebraic data type is instantiated, memory is allocated on the heap. A memory stack keeps track of the corresponding pointer to free the allocated memory. Memory deallocation is triggered on scope exit.