(platform-generic-c) Emulate return value for algebraic data types
Because of memory management for algebraic data types, instances of such types are deallocated on scope exit. A problem arises when the return value of a function happens to be an algebraic data type. This return value will be deallocated before the return statement. To avoid that, the return value is copied, the memory for the original return value is released, and the copy is returned instead.