(core, compiler, app-test) Change syntax of pattern expression
Pattern expression requires now parentheses surrounding the expression. A consequence of this change is the parsing of pattern wildcard, pattern binding and pattern deconstruction from the grammar itself rather than relying on a subsequent compilation phase. Since the gramamr now has production rules for all supported patterns, the pattern resolving phase only needs to resolve pattern bindings to pattern deconstructions if the underlying name refers to a type or a variant without fields.