(compiler) Add interpreter for constant expressions
The interpreter evaluates expressions whose value can be computed at compile time. To this end, it relies on symbolic values which are a direct mapping of values expressions can evaluate to. In addition, the interpreter uses a symbolic value - Undefined - denoting an expression whose value cannot be computed at compile time. The interpretation of an expression succeeds iff the evaluation is different from Undefined.