(core, compiler, platform-generic-c) Add support for metaprogramming
Metaprogramming applies to actors and type declarations. Actors and algebraic data types are now parametrizable with types and expressions. Parametrizing constructs with expressions requires such expressions to be evaluated at compile-time. The expression interpreter is used to this end. Parametrization is a two-step mechanism. First, parametrized expressions are turned into meta expressions. Then, meta expressions are instantiated with the provided type and value parameters. Instantiation is done through specialized substitution and loops until a substitution attemp does not modify the abstract syntax tree anymore. Specialized substitution means that there are as much declaration instances as there are declaration instantiations.