(compiler) Fix constant evaluator
The constant evaluator used to assume that integer literals are always unsigned numbers. The reason is due the way negative numbers are parsed; it is always an unary expression with minus as operation and the number as the operand. This needs to be relatex because the expression interpreter evaluates negative numbers to negative values.