(compiler) [tmp] Fix the assignability rules for numbers
The assignability rules used to raise typing errors for expressions mixing real and integer, and integers with different signedness or bit-length. This is no longer the case as expressions mixing integer and real should not lead to a compilation error. However, the convertibility rules, known now as the safe assignability rules are checked if the typing error was raised by the assignability rules. The main changes concern integer and real, and are summarized with the following rules:

A typing warning is raised when
- a 64-bit real is assigned to a 32-bit real,
- a real is assigned to an integer,
- an n-bit integer is assigned to an m-bit real if n > m,
- an integer is assigned to another integer with precision loss (sign or bit-length).

This probably needs to be refined. At least, there are no longer meaningless typing errors, only a few typing warning cornercases.